refactor:#140 체크리스트 진행도 컴포넌트 추가 및 로직분리
